+++ libs/core/select/select-mobile/select-mobile.module.ts
@@ -1,13 +1,11 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-
 import { DynamicComponentService } from '@fundamental-ngx/cdk/utils';
 import { ButtonComponent } from '@fundamental-ngx/core/button';
 import { DialogModule } from '@fundamental-ngx/core/dialog';
 import { SelectMobileComponent } from './select-mobile.component';
 
 @NgModule({
-    imports: [CommonModule, DialogModule, ButtonComponent, SelectMobileComponent],
+    imports: [DialogModule, ButtonComponent, SelectMobileComponent],
     exports: [SelectMobileComponent],
     providers: [DynamicComponentService]
 })