+++ libs/core/CHANGELOG.md
@@ -3,6 +3,180 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.55.0-rc.32](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.31...v0.55.0-rc.32) (2025-02-28)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.31](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.30...v0.55.0-rc.31) (2025-02-28)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.30](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.29...v0.55.0-rc.30) (2025-02-26)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.29](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.28...v0.55.0-rc.29) (2025-02-26)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.28](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.27...v0.55.0-rc.28) (2025-02-21)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.27](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.26...v0.55.0-rc.27) (2025-02-21)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.26](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.25...v0.55.0-rc.26) (2025-02-20)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.25](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.24...v0.55.0-rc.25) (2025-02-19)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.24](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.23...v0.55.0-rc.24) (2025-02-19)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.23](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.22...v0.55.0-rc.23) (2025-02-19)
+
+
+### Bug Fixes
+
+* **core:** removed up&down keys from tokens to fix rtl key handling ([#13044](https://github.com/SAP/fundamental-ngx/issues/13044)) ([52b25fa](https://github.com/SAP/fundamental-ngx/commit/52b25fa6d3993be425a5e79fa9afe68578333f68))
+
+
+
+
+
+# [0.55.0-rc.22](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.21...v0.55.0-rc.22) (2025-02-17)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.21](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.20...v0.55.0-rc.21) (2025-02-17)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.20](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.19...v0.55.0-rc.20) (2025-02-14)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.19](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.18...v0.55.0-rc.19) (2025-02-14)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.18](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.17...v0.55.0-rc.18) (2025-02-14)
+
+
+### Bug Fixes
+
+* **core:** illustrated messages responsiveness ([#13025](https://github.com/SAP/fundamental-ngx/issues/13025)) ([5ab8110](https://github.com/SAP/fundamental-ngx/commit/5ab81100fbd7f79604746012ace752148144992c))
+
+
+
+
+
+# [0.55.0-rc.17](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.16...v0.55.0-rc.17) (2025-02-13)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.16](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.15...v0.55.0-rc.16) (2025-02-13)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.15](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.14...v0.55.0-rc.15) (2025-02-13)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.14](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.13...v0.55.0-rc.14) (2025-02-13)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.13](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.12...v0.55.0-rc.13) (2025-02-13)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.55.0-rc.12](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.11...v0.55.0-rc.12) (2025-02-13)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
 # [0.55.0-rc.11](https://github.com/SAP/fundamental-ngx/compare/v0.55.0-rc.10...v0.55.0-rc.11) (2025-02-12)
 
 **Note:** Version bump only for package @fundamental-ngx/core